requirements.txt
boto3==1.5.6
decorator==4.1.2
Flask==0.12.2
futures==3.2.0
Jinja2==2.10
jmespath==0.9.3
PyYAML==3.12
